+++ doc/changelog.md
@@ -105,3 +105,5 @@
 - 2024/Jun/25: upgrade goutmp to 0.5.3; add aprilsh avatar.
 - 2024/Jun/30: fix diagnostics warning for souce; refine APK according to aports review (2nd review).
 - 2024/Jul/01: alpine aports approve aprilsh 0.7.5; thanks, Kevin Daudt.
+- 2024/Jul/15: still working on prediction engine.
+- 2024/Jul/17: overlay: add input echo mechanism to improve server side timeout.